Fix gosec errors
* Use `int32` for ports (Gateway API ports use that type)
* Use `int32` for timeouts, desired instances (pod healtcheck timeout is
  `int32`)
* Use `int64` for limits (k8s resources limits are `int64`

Thus we do not have to cast ints all over the place and fix gosec issues

+++ api/payloads/process.go
@@ -9,7 +9,7 @@ import (
 )
 
 type ProcessScale struct {
-	Instances *int   `json:"instances"`
+	Instances *int32 `json:"instances"`
 	MemoryMB  *int64 `json:"memory_in_mb"`
 	DiskMB    *int64 `json:"disk_in_mb"`
 }
@@ -34,9 +34,9 @@ type HealthCheck struct {
 }
 
 type Data struct {
-	Timeout           *int64  `json:"timeout"`
+	Timeout           *int32  `json:"timeout"`
 	Endpoint          *string `json:"endpoint"`
-	InvocationTimeout *int64  `json:"invocation_timeout"`
+	InvocationTimeout *int32  `json:"invocation_timeout"`
 }
 
 func (p ProcessScale) ToRecord() repositories.ProcessScaleValues {
